+++ Source/Map.cpp
@@ -472,7 +472,7 @@ void cMap::loadCF1Spt(tSharedBuffer pSpriteData, bool pCF2) {
 	uint16* SptPtr = (uint16*)pSpriteData->data();
 	uint16* SptFileEnd = SptPtr + (pSpriteData->size() / 2);
 
-	int32 Sprites = pSpriteData->size() / 8;
+	size_t Sprites = pSpriteData->size() / 8;
 	mSprites.resize(Sprites < g_Fodder->mParams.mSpritesMax ? g_Fodder->mParams.mSpritesMax : Sprites);
 	sSprite* Sprite = mSprites.data();
